Rumble: if all fails we will give you a star instead of crashing
In case you add a custom platform like rumble. You have to
supply icons. If you don't you get a weird crash that is not
so easy to see where it origintes from. This patch gives back
a valid resource instead. Maybe that is pull request worthy?
Maybe I try that! We will see!